move into separate file, check for extra deps
